refactor(web-shared): move Collapsible into ui directory
It's a generic UI primitive, not sidebar-specific.
